+++ packages/clients/snowbox/vite.config.js
@@ -1,3 +1,5 @@
 import { getClientConfig } from '../../../viteConfigs'
 
-export default getClientConfig()
+export default getClientConfig({
+  base: '',
+})