+++ packages/clients/meldemichel/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: This client now supports the `@polar/core`'s field `stylePath`. The usage is documented in the API.md file.
 - Feature: Update icon of `layerChooser` in `iconMenu` to `fa-layer-group` to clear-up the content hidden behind the menu button.
 - Feature: Update the close-button of the GFI window to indicate more clearly that it leads to the FeatureList.